# docs/swagger.yaml
basePath: /
host: localhost:4123
definitions:
...
  model.Environment:
    properties:
      config:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  types.PageResponse:
    properties:
      records: {}
      total:
        type: integer
    type: object
...
  types.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Minimal OpenAPI spec for Zebra-CICD endpoints
  title: Zebra-CICD API
  version: 0.1.0
paths:
...
  /api/environments:
    get:
      description: 获取所有环境的列表，支持按名称、类型、状态等条件查询
      parameters:
      - description: 环境名称
        in: query
        name: name
        type: string
      - description: 环境类型
        in: query
        name: type
        type: string
      - description: 环境状态
        in: query
        name: status
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 页数
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/types.PageResponse'
                  - properties:
                      records:
                        items:
                          $ref: '#/definitions/model.Environment'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取环境列表
      tags:
      - environments
    post:
      consumes:
      - application/json
      description: 创建一个新的环境
      parameters:
      - description: 环境信息
        in: body
        name: env
        required: true
        schema:
          $ref: '#/definitions/model.Environment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Environment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建环境
      tags:
      - environments
  /api/environments/{id}:
    delete:
      description: 根据环境ID删除环境（软删除）
      parameters:
      - description: 环境ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除环境
      tags:
      - environments
    get:
      description: 根据环境ID获取环境详情
      parameters:
      - description: 环境ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Environment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 根据ID获取环境
      tags:
      - environments
    put:
      consumes:
      - application/json
      description: 根据环境ID更新环境信息
      parameters:
      - description: 环境ID
        in: path
        name: id
        required: true
        type: integer
      - description: 环境信息
        in: body
        name: env
        required: true
        schema:
          $ref: '#/definitions/model.Environment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Environment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新环境
      tags:
      - environments
...
swagger: "2.0"